+++ integration_tests/handler/user_test.go
@@ -4,34 +4,16 @@ package handler_test
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"log"
 	"net/http"
-	"os"
-	"reflect"
-	"strconv"
 	"testing"
-	"time"
 
 	"github.com/gofrs/uuid"
-	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
-	"github.com/traPtitech/traPortfolio/infrastructure"
 	"github.com/traPtitech/traPortfolio/interfaces/handler"
 	"github.com/traPtitech/traPortfolio/usecases/repository"
 )
 
 var (
-	port = func() int {
-		if p, err := strconv.Atoi(os.Getenv("PORT")); err == nil {
-			return p
-		}
-
-		return 1323
-	}()
-	baseURL = fmt.Sprintf("http://localhost:%d/api/v1", port)
-
 	sampleUser1 = User{
 		Id:       uuid.FromStringOrNil("11111111-1111-1111-1111-111111111111"),
 		Name:     "user1",
@@ -48,76 +30,6 @@ var (
 		RealName: "東 工子",
 	}
 )
-
-func mustMarshal(v interface{}) []byte {
-	b, err := json.Marshal(v)
-	if err != nil {
-		panic(err)
-	}
-
-	return b
-}
-
-func TestMain(m *testing.M) {
-	s := infrastructure.NewSQLConfig("root", "password", "localhost", "portfolio", 3307)
-	t := infrastructure.NewTraQConfig("", "", true)
-	p := infrastructure.NewPortalConfig("", "", true)
-	k := infrastructure.NewKnoqConfig("", "", true)
-	api, err := infrastructure.InjectAPIServer(&s, &t, &p, &k)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	go func(api handler.API) {
-		e := echo.New()
-		handler.Setup(e, api)
-
-		log.Fatal(e.Start(fmt.Sprintf(":%d", port)))
-	}(api)
-
-	for i := 0; ; i++ {
-		log.Printf("waiting for server to start... (%ds)", i)
-
-		if i > 10 {
-			log.Fatal("failed to connect to server")
-		}
-
-		if _, err := http.Get(baseURL + "/ping"); err == nil {
-			break
-		}
-
-		time.Sleep(time.Second)
-	}
-
-	m.Run()
-}
-
-// GET /ping
-func TestPing(t *testing.T) {
-	t.Parallel()
-	tests := map[string]struct {
-		statusCode int
-		want       []byte
-	}{
-		"200": {http.StatusOK, []byte("pong")},
-	}
-	for name, tt := range tests {
-		tt := tt
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			cli, err := NewClientWithResponses(baseURL)
-			assert.NoError(t, err)
-
-			res, err := cli.PingWithResponse(context.Background())
-			assert.NoError(t, err)
-
-			assert.Equal(t, tt.statusCode, res.StatusCode())
-
-			actual := reflect.ValueOf(*res).FieldByName("Body").Interface().([]byte)
-			assert.Equal(t, tt.want, actual)
-		})
-	}
-}
 
 // GET /users
 func TestGetUsers(t *testing.T) {
